# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de la plateforme QUI PERD (défis de match, escrow, paiements Mobile Money).",
        "title": "QUI PERD API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/defis": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "defis"
                ],
                "summary": "Défis ouverts (?jeu=, ?plateforme=, ?miseMax=) — ou mes défis, tous statuts (?mes=1)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/defis.DefiListe"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "defis"
                ],
                "summary": "Créer un défi (bloque la mise du créateur)",
                "responses": {}
            }
        },
...
        "/defis/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "defis"
                ],
                "summary": "Détail d'un défi (libellés créateur/jeu/plateforme) et son match s'il existe",
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "defis"
                ],
                "summary": "Annuler un défi ouvert (rend la mise au créateur, moins la commission)",
                "responses": {}
            }
        },
        "/defis/{id}/rejoindre": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "defis"
                ],
                "summary": "Rejoindre un défi (bloque la mise, crée le match)",
                "responses": {}
            }
        },
...
        "/matchs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "matchs"
                ],
                "summary": "Mes matchs (tableau, ?statut=en_cours|verification|litige|termine) ; admin ?tous=1 : page utils.Page[MatchEnrichi] (10/page, ?page\u0026taille)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "en_cours | verification | litige | termine",
                        "name": "statut",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Admin : 1 = tous les matchs, réponse paginée",
                        "name": "tous",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Admin (?tous=1) : page, défaut 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Admin (?tous=1) : éléments par page, 1..100, défaut 10",
                        "name": "taille",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/matchs.MatchEnrichi"
                            }
                        }
                    }
                }
            }
        },
        "/matchs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "matchs"
                ],
                "summary": "Détail d'un match (libellés joueurs/jeu/plateforme + déclarations)",
                "responses": {}
            }
        },
        "/matchs/{id}/declaration": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "matchs"
                ],
                "summary": "Déclarer le score d'un match",
                "responses": {}
            }
        },
...
        "/matchs/{id}/validation": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "matchs"
                ],
                "summary": "Valider un match (admin) — déclenche le règlement de l'escrow",
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "defis.DefiListe": {
            "type": "object",
            "properties": {
                "createurId": {
                    "type": "string"
                },
                "createurNom": {
                    "type": "string"
                },
                "dateCreation": {
                    "type": "string"
                },
                "dateExpiration": {
                    "type": "string"
                },
                "dateModification": {
                    "type": "string"
                },
                "devise": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jeuCategorie": {
                    "type": "string"
                },
                "jeuId": {
                    "type": "string"
                },
                "jeuNom": {
                    "type": "string"
                },
                "montantMise": {
                    "type": "number"
                },
                "plateformeFamille": {
                    "type": "string"
                },
                "plateformeId": {
                    "type": "string"
                },
                "plateformeNom": {
                    "type": "string"
                },
                "regles": {
                    "type": "string"
                },
                "statut": {
                    "type": "string"
                }
            }
        },
...
        "matchs.MatchEnrichi": {
            "type": "object",
            "properties": {
                "dateCreation": {
                    "type": "string"
                },
                "dateDebut": {
                    "type": "string"
                },
                "dateFin": {
                    "type": "string"
                },
                "defiId": {
                    "type": "string"
                },
                "devise": {
                    "type": "string"
                },
                "gagnantId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jeuNom": {
                    "type": "string"
                },
                "joueur1Id": {
                    "description": "Noms de colonnes explicites : GORM dériverait `joueur1_id` / `score_joueur1`,\nalors que le modèle de données (§7, table 6) impose `joueur_1_id` / `score_joueur_1`,\nnoms utilisés par les requêtes SQL brutes (litiges, scores).",
                    "type": "string"
                },
                "joueur1Nom": {
                    "type": "string"
                },
                "joueur2Id": {
                    "type": "string"
                },
                "joueur2Nom": {
                    "type": "string"
                },
                "montantMise": {
                    "type": "number"
                },
                "perdantId": {
                    "type": "string"
                },
                "plateformeNom": {
                    "type": "string"
                },
                "scoreJoueur1": {
                    "type": "integer"
                },
                "scoreJoueur2": {
                    "type": "integer"
                },
                "statut": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}